ARM: imx6q-evi: Make standalone dts for rev xf
Rev xf boards have no spi3 and several differences on ecspi5
there are 3 working devices on spi5.

There are 2 adc/temp sensors also connected, but their voltage
sensors are not wired, causing the instrument to believe there
is a short, so they are omitted.
There is also a  6th device, an RTC which cannot work due to
incorrect wiring. This will not be added to the dts