misc: evi-eddy: Remove all references to hw type
Remove EIM or FPGA from all names, including the device file,
which goes from "evifpga" to "evi-eddy"
This is a backwards incompatible change, as we prepare to move the driver
out of the evi repository.

The driver should describe the hardware it is driving,
not the implementation on the other side of the ABI.

The fact that it is implemented in an FPGA
is in no way germaine to the function of the driver, which
will be the same if/when it is in an ASIC.

We are working on a pcie variant, so EIM is no longer
appropriate here.